chore: directly generate gql types from remote schema
