+++ apps/publikator/graphql/configs/codegen-dev.ts
@@ -8,7 +8,7 @@ const config: CodegenConfig = {
   overwrite: true,
   generates: {
     './graphql/republik-api/__generated__/gql/': {
-      schema: './graphql/republik-api/__generated__/schema.graphql',
+      schema: process.env.NEXT_PUBLIC_API_URL,
       documents: ['./graphql/republik-api/**/*.{ts,tsx,gql,graphql}'],
       preset: 'client',
       presetConfig: {
