+++ apps/publikator/package.json
@@ -8,15 +8,14 @@
   "description": "",
   "main": "server/index.js",
   "scripts": {
-    "dev": "yarn gql:schema && concurrently \"PORT=3005 next dev\" \"yarn gql:types --watch\"",
+    "dev": "concurrently \"PORT=3005 next dev\" \"yarn gql:types --watch\"",
     "start": "next start",
-    "build": "yarn gql:schema && yarn gql:types && next build",
+    "build": "yarn gql:types && next build",
     "test": "jest",
     "cloc": "find . -name '*.js' -not -path \"./node_modules*\" -not -path \"./.next*\" | xargs wc -l",
     "lint": "eslint --fix .",
     "analyze": "ANALYZE=true next build",
-    "gql:types": "graphql-codegen --config graphql/configs/codegen-dev.ts",
-    "gql:schema": "graphql-codegen --config graphql/configs/codegen-schema.ts"
+    "gql:types": "graphql-codegen --config graphql/configs/codegen-dev.ts"
   },
   "repository": {
     "type": "git",
