+++ apps/www/graphql/configs/codegen-dev.ts
@@ -4,11 +4,33 @@ import { loadEnvConfig } from '@next/env'
 
 loadEnvConfig(process.cwd())
 
+if (!process.env.DATO_CMS_API_URL) {
+  throw new Error('Please set DATOCMS_API_URL in your .env file')
+}
+if (!process.env.DATO_CMS_API_TOKEN) {
+  throw new Error('Please set DATOCMS_API_TOKEN in your .env file')
+}
+
+const headers = {
+  Authorization: process.env.DATO_CMS_API_TOKEN,
+  'X-Exclude-Invalid': 'true',
+}
+
+if (process.env.DATO_CMS_ENVIRONMENT) {
+  headers['X-Environment'] = process.env.DATO_CMS_ENVIRONMENT
+}
+
 const config: CodegenConfig = {
   overwrite: true,
   generates: {
     './graphql/cms/__generated__/gql/': {
-      schema: './graphql/cms/__generated__/schema.graphql',
+      schema: [
+        {
+          [process.env.DATO_CMS_API_URL]: {
+            headers,
+          },
+        },
+      ],
       documents: ['./graphql/cms/**/*.{ts,tsx,gql,graphql}'],
       preset: 'client',
       presetConfig: {
@@ -26,7 +48,7 @@ const config: CodegenConfig = {
       plugins: [],
     },
     './graphql/republik-api/__generated__/gql/': {
-      schema: './graphql/republik-api/__generated__/schema.graphql',
+      schema: process.env.NEXT_PUBLIC_API_URL,
       documents: ['./graphql/republik-api/**/*.{ts,tsx,gql,graphql}'],
       preset: 'client',
       presetConfig: {
